requirements.txt
numpy>=1.24.0
matplotlib>=3.7.0
scikit-learn>=1.3.0
rerun-sdk>=0.15.0
torch>=2.0.0
hydra-core>=1.3.0
tqdm>=4.65.0
